pyproject.toml
[project]
name = "cv-coverletter-generator"
version = "0.1.0"
description = "AI-powered CV/Resume and Cover Letter Generator using Streamlit and Gemini API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.31.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]

[tool.uv]
dev-dependencies = []
